Fix compiler warnings and convert value to uintmax_t for "%ju" format string.
Submitted by: Serge Vakulenko

git-svn-id: svn://svn.code.sf.net/p/elftoolchain/code/trunk@3277 95820547-d848-0410-985e-9ae8fe0fa350

+++ ld/amd64.c
@@ -133,7 +133,7 @@ _reloc2str(uint64_t r)
 		case 22: return "R_X86_64_GOTTPOFF";
 		case 23: return "R_X86_64_TPOFF32";
 	default:
-		snprintf(s, sizeof(s), "<unkown: %ju>", r);
+		snprintf(s, sizeof(s), "<unkown: %ju>", (uintmax_t) r);
 		return (s);
 	}
 }
